Sunny Kamboz Indian Singer
Sunny Kamboz is an Indian Singer | Songwriter | Composser | Music Produce. He was born on 17 March 1994 in Patiala. He is the active ceo and founder of "SK Music Studio", a digital marketing company. The best way to keep up is to learn from and support each other as we continue pushing the boundaries of innovation in marketing.
He is a graduate ( B.A. Arts) student at Government Mohindra College.
Sunny Kamboz is an Influencer, fashion blogger, and Social media star. Sunny Kamboz is yet unmarried. He also has not posted any picture of her girlfriend in the social media. Therefore, we can believe that the star is still single.